refactor: move page into components
