+++ src/app/pages/home/Home.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-import { HeroBanner } from "@app/pages/home/container/HeroBanner";
-import { FriendBanner } from "@app/pages/home/container/FriendBanner";
-import { Sellers } from "@app/pages/home/container/Sellers";
-import { Adoption } from "@app/pages/home/container/Adoption";
+import { HeroBanner } from "@app/pages/home/components/HeroBanner";
+import { FriendBanner } from "@app/pages/home/components/FriendBanner";
+import { Sellers } from "@app/pages/home/components/Sellers";
+import { Adoption } from "@app/pages/home/components/Adoption";
 
 export const Home = () => {
   return (
