+++ src/app/pages/home/components/Adoption.tsx
