+++ src/app/pages/home/components/FriendBanner.tsx
