+++ src/app/pages/home/components/HeroBanner.tsx
